[WIP] feat: various - TLS configuration for pcap-bosh-cli against BoshDirectorAPI & pcap-api - refactor client - [WIP] unit tests for client - minor type change in pcap.proto - pb type fix for pcap-api-cli

+++ src/pcap/pcap.go
@@ -85,7 +85,7 @@ func newPacketResponse(data []byte, captureInfo gopacket.CaptureInfo) *CaptureRe
 			Packet: &Packet{
 				Data:      data,
 				Timestamp: timestamppb.New(captureInfo.Timestamp),
-				Length:    uint64(captureInfo.Length),
+				Length:    int32(captureInfo.Length),
 			},
 		},
 	}
